refactor(web): extract VideoPlayer controller into a hook
Move the media state, effects, and imperative handlers into a
useVideoPlayer hook and the control bar into VideoControlBar, leaving
VideoPlayer as a presentational shell.
